Add messages system and game state
Added a messages system to properly handle conditional messages in the game loop print statement.  For example not every sector has a port so the port should only be displayed when a port exists.  The message system is a list that message lines are appended to and the get_messages function contains all of the conditional logic for when messages should appear and how they should be formatted.

Added an initial implementation of game state.  The game state is saved as the graph attribute 'state' on the Universe graph.  Currently the game state maintains the name, location, wallet and visit history of each player.