Add initial port interaction
Move bimodal function from bigbang to utils

Created components.py to hold classes

Created the commodity component class to generate commodity items and prices for stations.

Improved the prints in bigbang.py to more clearly communicate what is happening in the Universe creation process.

Modified the stationgen function to utilize the new Commodity class to create the station items.  The idea behind the Commodity class is to make it so that the items are not 'hard coded' and new items can be easily added to new games and things will just work with the trading system.  The remaining issue is that the station generation process attempts to create a set of stations that provide opportunities for trade.  In order to generate a playable set of stations I generate stations with commodity prices pulled from a random distribution.  I then evaluate the station and apply certain tags to it.  For example a station with a very low 'ice' sale price is flagged as an 'ICE MINING' station. I attempt to generate a set of stations with at least a few resource producing stations, for example the 'ICE MINING' tag, and a few production stations with high buy prices for Ice and other resources.  In order to do this I currently have the tagging process use 'hard coded' commodities.  It is probably possible to abstract this idea to enable the tagging process without 'hard coding' the commodities but the commodity class or some other structure would need to describe the natural relationships between commodities and how they might be utilized or traded within the game Universe.

Addes holds and cargo to player dict to facilitate trading.  The player dict is getting larger and larger and there will need to be code to evaluate the cargohold and other things for trading.  Questions about where this code will live need to be resolved.  One option would be to create a Player class to handle some of these things.

Added the P command to allow players to interact with ports if a system has a port.  The T option on the port menu allows for trading with a port.  Currently the option only prints the commodities, prices and volumes offered by the station.