complete initial prototype
Bigbang now places ice mining stations on Ice and Water planets instead of in sectors. Non-ice mining stations are only placed in sectors

Fixed off by one error in bodies.  Going to 14.1 for example was actually taking you to 14.2 due to the slicing

Fixed the Port and Trade menu to accept t as well as T to launch trading

Updated the help dialog

Added id field to stars and bodies, id is uuid generated when the star/body are generated

Add scanner to show star and planet data